Better queue length estimate
Assuming we have a throughput of 1 job every 2 minutes (calculated
from our capacity of 60 and an average job length of 2 hours), it
should be possible to get a rough estimate of how long it will take
a given job to get through the queue.  This is also a more useful
metric for the overall stats, so use that number instead of the
wait time of the first job.