Make estimated time smarter
It's nonsense to have an estimated time on a completed job, and
once a job is running the estimated time isn't based on the queue
depth but on how long it has been running.  Fixing the latter also
means the estimates will be somewhat more useful when the queue is
not full too.