Revert back to zuulv2
The zuulv3 migration has been partially rolled back, which means the
relevant jobs are running in zuulv2 now.

Note that the nodepool file is still the zuulv3 one, but that only
messes up the max jobs count which is not a big deal.  I'm not
inclined to revert those changes at the moment.