Add quick and dirty all queue functionality
This allows a way to list all changes from all recognized queues in
a given zuul instance.  Currently this is pretty basic, in that it
just dumps all of the changes into one list and doesn't have any
display to indicate which queue a change came from.  It is mostly
intended to allow filtering across all queues.