Improve error reporting
Previously if the data we got back from zuul was not what we expected,
the user just got an internal server error message.  Since we
already have some error handling code, we can provide a little bit
more information and a nicer looking error page by also handling
problems processing the data.

The error is also reported as the repr() instead of str() because
otherwise things like KeyErrors stringify to just the missing key,
which is not terribly helpful.